+++ src/discrete-uniform-distribution.cpp
@@ -30,7 +30,7 @@ using Rcpp::NumericMatrix;
 
 
 double pmf_dunif(double x, double min, double max) {
-  if (min >= max || std::isinf(min) || std::isinf(max) ||
+  if (min > max || std::isinf(min) || std::isinf(max) ||
       floor(min) != min || floor(max) != max) {
     Rcpp::warning("NaNs produced");
     return NAN;
@@ -42,7 +42,7 @@ double pmf_dunif(double x, double min, double max) {
 
 
 double cdf_dunif(double x, double min, double max) {
-  if (min >= max || std::isinf(min) || std::isinf(max) ||
+  if (min > max || std::isinf(min) || std::isinf(max) ||
       floor(min) != min || floor(max) != max) {
     Rcpp::warning("NaNs produced");
     return NAN;
@@ -55,7 +55,7 @@ double cdf_dunif(double x, double min, double max) {
 }
 
 double invcdf_dunif(double p, double min, double max) {
-  if (min >= max || std::isinf(min) || std::isinf(max) ||
+  if (min > max || std::isinf(min) || std::isinf(max) ||
       floor(min) != min || floor(max) != max) {
     Rcpp::warning("NaNs produced");
     return NAN;